backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==23.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
exceptiongroup==1.2.2
fastapi==0.115.3
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httplib2==0.14.0
huggingface-hub==0.26.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
joblib==1.4.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==5.3.0
MarkupSafe==1.1.0
more-itertools==4.2.0
mpmath==1.3.0
netifaces==0.10.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
packaging==24.1
pexpect==4.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydantic==2.9.2
pydantic-core==2.23.4
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMuPDF==1.24.11
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pypdf2==3.0.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-debian===0.1.36ubuntu1
python-docx==1.1.2
python-multipart==0.0.12
PyYAML==5.3.1
regex==2024.9.11
requests==2.22.0
requests-unixsocket==0.2.0
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==2.3.1
sentencepiece==0.2.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
sos==4.4
ssh-import-id==5.10
starlette==0.41.0
sympy==1.13.3
systemd-python==234
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.4.1
tqdm==4.66.6
transformers==4.46.1
triton==3.0.0
Twisted==18.9.0
typing-extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.32.0
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
